+++ Assets/Scripts/createImage.js
